# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server docs",
        "title": "Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
        "/v1/profile": {
            "get": {
                "description": "admin profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo_user.UserProfileResDto"
                        }
                    }
                }
            }
        },
        "/v1/sign-in": {
            "post": {
                "description": "sign in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign"
                ],
                "parameters": [
                    {
                        "description": "sign in body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo_user.SignInReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ResDto"
                        }
                    }
                }
            }
        },
        "/v1/sign-out": {
            "post": {
                "description": "sign out user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ResDto"
                        }
                    }
                }
            }
        },
        "/v1/sign-up": {
            "post": {
                "description": "sign up user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign"
                ],
                "parameters": [
                    {
                        "description": "sign up body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo_user.SignUpReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.ResDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "config.ErrDto": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "config.ResDto": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.ErrDto"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "repo_user.SignInReqDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "repo_user.SignUpReqDto": {
            "type": "object",
            "properties": {
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "repo_user.UserProfileDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "repo_user.UserProfileResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/repo_user.UserProfileDto"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.ErrDto"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}